pyproject.toml
[tool.poetry]
name = "financeapp"
version = "0.1.0"
description = ""
authors = ["Ekaterina Itchenko <ekaterina.savickaya.1989@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.8.19"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
djoser = "^2.2.0"
whitenoise = "^6.4.0"
django = "^4.2.1"
gunicorn = "^20.1.0"
drf-yasg = "^1.21.5"
django-cors-headers = "^4.0.0"
types-requests = "2.31.0.20240406"
requests = "^2.31.0"
drf-spectacular = "0.26.5"
drf-spectacular-sidecar = "2023.10.1"
psycopg2-binary = "2.9.9"
rich = "^13.0"
django-filter = "^24.3"
django-stubs = "4.2.7"
djangorestframework-stubs = {version = "1.7.0", extras = ["compatible-mypy"]}

[tool.poetry.group.dev.dependencies]
coverage = "^7.2.6"
rich = "^13.5.2"
setuptools = "*"
pre-commit = "3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
